Fix various casting warnings
git-svn-id: https://lmms.svn.sf.net/svnroot/lmms/trunk/lmms@1471 0778d3d1-df1d-0410-868b-ea421aaaa00d

+++ src/gui/automation_editor.cpp
@@ -642,7 +642,7 @@ void automationEditor::leaveEvent( QEvent * _e )
 
 void automationEditor::drawLine( int _x0, float _y0, int _x1, float _y1 )
 {
-	int deltax = tAbs<float>( _x1 - _x0 );
+	int deltax = static_cast<int>( tAbs<float>( _x1 - _x0 ) );
 	float deltay = tAbs<float>( _y1 - _y0 );
 	int x = _x0;
 	float y = _y0;
